requirements.txt
absl-py==1.4.0
anytree==2.8.0
certifi==2022.12.7
h5py
imagecodecs
imageio
imagesize
imbalanced-learn
importlib-metadata
incremental
inflection==0.5.1
iniconfig
intake
intervaltree
ipykernel
ipython
json5 
jsonpatch
keras==2.11.0
lxml
matplotlib
multipledispatch
nltk 
numpy
opencv-python==4.7.0.72
openpyxl==3.0.10
pandas
Pillow==9.4.0
poyo
protobuf==4.23.3
psutil 
ptyprocess
pycocotools==2.0.6
pycosat
PyYAML
scikit-image
scikit-learn 
scipy==1.10.0
six
tables==3.7.0
tabulate @ file:///C:/ci/tabulate_1657600805799/work
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tf-slim==1.1.0
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tornado
tqdm 